feat: Add MediatR and Application layer
Added MediatR package to the Api project and created a new Application layer. The Application project references the Domain layer and MediatR.Contracts package, while the Infrastructure project was updated to reference the Application project. MediatR was also configured in ConfigureServices with services from the Assembly housing the Application layer.