feat: Add Result classes and corresponding unit tests
Added new Result and Result<T> classes in the Application.Results namespace to represent the result of various operations in the application. These classes carry success or failure status, as well as a list of exceptions if any. Corresponding unit tests are also added to ensure proper functioning of these classes. The Application.Tests project has been duly updated and added to the solution file.